requirements.txt
articlemetaapi==1.6.13